day-14/pyproject.toml
[project]
name = "day-14"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
]
